Deploy script complete
- Completed deploy script
- Tested with a WP install on Vultr server
- Should work for any server where the wordpress installation sits at
/var/www/html and is owned by www-data.
- Implementing this to work for a variable directory and user should be
trivial
- Could check if the user is already www-data as well (fairly narrow
scope at the moment)

+++ bin/deploy.sh
@@ -1,14 +1,24 @@
 #!/usr/bin/env bash
-# INCOMPLETE
+
 # Script to deploy a wordpress theme to a server.
 # Currently based on just one theme, with the standard WP folder layout,
 # sitting at /var/www/html.
+# Assumes web user is www-data
+
+# LIVE_URL & DEV_URL variables need to be set in .env one directory up from
+# the script's location.
 
 if [ $# != 3 ] 
 then
     echo -e 'Incorrect number of arguments\nUsage: deploy <username> <address> <theme name>'
     exit 1
 fi
+
+DEV_URL=$(grep -oP '^DEV_URL=\K.*' .env)
+LIVE_URL=$(grep -oP '^LIVE_URL=\K.*' .env)
+
+[ -z $DEV_URL ] && { echo "DEV_URL not set in .env, exiting ..."; exit 1; }
+[ -z $LIVE_URL ] && { echo "LIVE_URL not set in .env, exiting ..."; exit 1; }
 
 temp_dir=DEPLOY_TEMP;
 cd $(dirname $0)/..;
@@ -25,6 +35,7 @@ mv plugins.tar.gz $start_dir/$temp_dir/;
 # Get the theme
 echo 'Getting theme ...';
 cd ../themes;
+[ -d $3 ] || { echo "$3 theme doesn't exist, exiting ..."; rm -rf $start_dir/$temp_dir; exit 1; }
 tar -zcf $3.tar.gz $3;
 mv $3.tar.gz $start_dir/$temp_dir/;
 
@@ -34,16 +45,46 @@ cd ../uploads;
 tar -zcf uploads.tar.gz *;
 mv uploads.tar.gz $start_dir/$temp_dir/;
 
-cd $start_dir/$temp_dir;
+echo 'Getting the database dump ...'
+cp $start_dir/data/dump.sql $start_dir/$temp_dir
+
+cd $start_dir;
 echo 'Transferring files to server ...';
-scp * $1@$2:/var/www;
+scp -r $temp_dir $1@$2:~;
 
-echo 'Doing stuff with the files ...';
-# TODO: Continue here
-ssh -t $1@$2 'cd /var/www; \
-touch i_was_here';
+config_parse() {
+    grep -xoP "define\(\s?['\`\"]$1['\`\"][^'\`\"]*['\`\"]\K([^'\`\"]+)['\`\"]\s?\);" wp-config.php |\
+    sed -r "s/['\`\\\"]/'/g" | cut -d "'" -f 1
+}
 
-echo 'Cleaning up ...';
+ssh -t $1@$2 "$(declare -f config_parse); \
+cd /var/www/html; \
+DB_USER=\$(config_parse DB_USER); \
+DB_NAME=\$(config_parse DB_NAME); \
+DB_PASSWORD=\$(config_parse DB_PASSWORD); \
+echo 'Importing the DB ...'; \
+mysql -u \$DB_USER -p\$DB_PASSWORD \$DB_NAME < ~/$temp_dir/dump.sql; \
+echo 'Updating the DB ...'; \
+wp --allow-root search-replace $DEV_URL $LIVE_URL --skip-columns=guid; \
+echo 'Unpacking the plugins ...'; \
+cd wp-content/plugins; \
+mv ~/$temp_dir/plugins.tar.gz ./; \
+tar -zxf plugins.tar.gz; \
+echo 'Unpacking uploads ...'; \
+cd ../uploads; \
+mv ~/$temp_dir/uploads.tar.gz .; \
+tar -zxf uploads.tar.gz; \
+echo 'Unpacking the theme ...'; \
+cd ../themes; \
+mv ~/$temp_dir/$3.tar.gz .; \
+tar -zxf $3.tar.gz; \
+echo 'Cleaning up on server ...'; \
+cd /var/www; \
+chown -R www-data:www-data html; \
+chown \$USER:\$USER html/wp-config.php; \
+rm -rf ~/$temp_dir;"
+
+echo 'Cleaning up temporary files ...';
 rm -rf $start_dir/$temp_dir;
 
-echo 'Done';
+echo 'Done!';
